requirements.txt
fastmcp
uvicorn
requests
python-dotenv
psycopg2
psycopg2-binary
azure-storage-blob
snowflake-connector-python
psycopg2-binary
cryptography